requirements.txt
tkcalendar
openpyxl
Pillow
pywin32
bcrypt
mysql-connector-python
